Pesquisa por categoria (em andamento)

+++ src/br/livraria/dao/PesquisaDao.java
@@ -11,6 +11,6 @@ public interface PesquisaDao {
 	
 	public List<Livro> pesquisarPorAutor(Integer codigoAutor) throws SQLException;
 	public List<Livro> pesquisarPorTitulo(String titulo) throws SQLException;
-	public List<Livro> pesquisarPorCategoria(Integer codigoCategoria) throws SQLException;
+	public List<Livro> pesquisarPorCategoria(Long codigoCategoria) throws SQLException;
 	
 }
